bf: get bool array in cs, marshaling issue

+++ EngineEx_Template.cs
@@ -357,7 +357,7 @@ namespace NAMESPACE_NAME
 
             if (card > 0)
             {
-                var values = new Int64[card];
+                var values = new byte[card];
                 System.Runtime.InteropServices.Marshal.Copy(valuesPtr, values, 0, (int)card);
 
                 var bools = new bool[card];
